Implemented first version of "redeem voucher"

+++ src/main/java/com/senacor/bankathon2018/webendpoint/model/returnDTO/BoughtVoucherDTO.java
@@ -6,7 +6,7 @@ public class BoughtVoucherDTO {
 
   private Long id;
 
-  private Long voucherId;
+  private Integer voucherId;
 
   private String voucherName;
 
@@ -24,11 +24,11 @@ public class BoughtVoucherDTO {
     this.id = id;
   }
 
-  public Long getVoucherId() {
+  public Integer getVoucherId() {
     return voucherId;
   }
 
-  public void setVoucherId(Long voucherId) {
+  public void setVoucherId(Integer voucherId) {
     this.voucherId = voucherId;
   }
